requirements.txt
flask==2.0.1
Pillow==8.3.1
werkzeug==2.0.1
